LibWeb: Split InitialStyleValue out of StyleValue.{h,cpp}
