+++ Userland/Libraries/LibWeb/CSS/StyleValues/InitialStyleValue.h
@@ -1,0 +1,33 @@
+/*
+ * Copyright (c) 2022-2023, Sam Atkins <atkinssj@serenityos.org>
+ *
+ * SPDX-License-Identifier: BSD-2-Clause
+ */
+
+#pragma once
+
+#include <LibWeb/CSS/StyleValue.h>
+
+namespace Web::CSS {
+
+class InitialStyleValue final : public StyleValueWithDefaultOperators<InitialStyleValue> {
+public:
+    static ValueComparingNonnullRefPtr<InitialStyleValue> the()
+    {
+        static ValueComparingNonnullRefPtr<InitialStyleValue> instance = adopt_ref(*new InitialStyleValue);
+        return instance;
+    }
+    virtual ~InitialStyleValue() override = default;
+
+    ErrorOr<String> to_string() const override { return "initial"_string; }
+
+    bool properties_equal(InitialStyleValue const&) const { return true; }
+
+private:
+    InitialStyleValue()
+        : StyleValueWithDefaultOperators(Type::Initial)
+    {
+    }
+};
+
+}
